requirements.txt
annotated-types==0.6.0
anyio==4.2.0
black==24.1.1
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.109.0
flake8==7.0.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
Mako==1.3.2
MarkupSafe==2.1.4
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pydantic==2.6.0
pydantic_core==2.16.1
pyflakes==3.2.0
PyJWT==2.8.0
python-decouple==3.8
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
tomli==2.0.1
typing_extensions==4.9.0
uvicorn==0.27.0.post1
